added a file to select VM size and image
A script was created which reads the available VM data and compares it with the selected regions data to see if a VM is present in the regions selected. If it is then the script asks for confirmation and stores the data into a file. The next step will be to read the file that will be created by the selector and base the deployment on that.